+++ src/Hades.Syntax/Expression/Nodes/VariableDeclarationNode.cs
@@ -19,6 +19,7 @@ namespace Hades.Syntax.Expression.Nodes
         public Node Assignment { get; set; }
         public bool Dynamic { get; set; }
         public bool Nullable { get; set; }
+        public string SpecificType { get; set; }
 
         protected override string ToStr()
         {
@@ -28,6 +29,12 @@ namespace Hades.Syntax.Expression.Nodes
             var array = Array ? " array" : "";
             var arraySize = ArraySize != null ? " (" + ArraySize + ")" : "";
             var datatype = Datatype != null ? " " + Datatype.ToString().ToLower() : "";
+
+            if (SpecificType != null)
+            {
+                datatype += $"[{SpecificType}]";
+            }
+            
             var assignment = Assignment != null ? "(" + Assignment + ")" : "";
             var withAssignment = assignment != "" ? " with assignment " : "";
             return $"Create{nullable}{dynamic}{mutable}{datatype}{array}{arraySize} variable {Name}{withAssignment}{assignment}";