+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 4/26/19
+
+### Added specific object and proto validation to lambdas, function and variables #6623878e8ccec87a20c395cdd157931dea2b403b
+
+* Variable instantiations, function arguments and lambda arguments can now name specific class-/proto names
+
+  ```swift
+  func doStuff(args object::IClient a)
+  	a->stuff("Test")
+  end
+  ```
+
+  ```swift
+  var x = {args object::IClient a, int b =>
+  	a->moreStuff(b)
+  }
+  ```
+
+  ```js
+  var proto::console c
+  ```
+
 ## 4/23/19
 
 ### Added lambda parameter types #c244d04d419cb163ce7e5581990df1dd39c44190